second pass refactor
