+++ components/Dialogs/ConfigDialog/index.ts
