+++ components/Dialogs/index.ts
@@ -1,0 +1,2 @@
+export * from './AppDialog';
+export * from './ConfigDialog';
