+++ next.config.js
@@ -2,6 +2,7 @@
 const withPWA = require('next-pwa')({
   dest: 'public',
   disable: process.env.NODE_ENV === 'development',
+  l,
 });
 
 module.exports = withPWA({
